Refactor the download package
